refactor: remove @Inject for constructor
