--- app/src/main/res/layout/fragment_lottie.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-	xmlns:tools="http://schemas.android.com/tools"
-	android:layout_width="match_parent"
-	android:layout_height="match_parent"
-	xmlns:app="http://schemas.android.com/apk/res-auto"
-	tools:context=".ui.lottie.LottieFragment">
-
-	<com.airbnb.lottie.LottieAnimationView
-		android:id="@+id/animation"
-		android:layout_width="match_parent"
-		android:layout_height="match_parent"
-		android:scaleType="fitXY"
-		app:lottie_loop="false"
-		app:lottie_fileName="lottie.json" />
-</FrameLayout>
