+++ app/src/main/res/layout/fragment_splash_screen.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+
+<com.airbnb.lottie.LottieAnimationView
+	xmlns:android="http://schemas.android.com/apk/res/android"
+	xmlns:app="http://schemas.android.com/apk/res-auto"
+	android:id="@+id/animation"
+	android:layout_width="match_parent"
+	android:layout_height="match_parent"
+	android:scaleType="fitXY"
+	app:lottie_loop="false"
+	app:lottie_fileName="lottie.json" />
